pyproject.toml
[tool.poetry]
name = "."
version = "0.1.0"
description = ""
authors = ["jacksonsr451 <jackson.severino.rocha@escola.pr.gov.br>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
django-role-permissions = "^3.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
